requirements.txt
streamlit==1.36.0
pandas==2.2.0
yfinance
numpy
Pillow
scikit-learn
tensorflow
matplotlib


